pyproject.toml
[tool.poetry]
name = "no-country-equipo-22"
version = "0.1.0"
description = ""
authors = ["eseminara <eze.seminara@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
pdfplumber = "^0.11.7"
python-docx = "^1.2.0"
pytesseract = "^0.3.13"
pillow = "^11.3.0"
dotenv = "^0.9.9"
django-cors-headers = "^4.9.0"
django-filter = "^25.1"
psycopg2-binary = "^2.9.10"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
